🎨 style: コンポーネントを分割
