+++ frontend/src/feature/users/components/UserEdit/index.ts
@@ -1,0 +1,1 @@
+export { UserEdit } from './UserEdit'
