+++ frontend/src/feature/users/components/UserEditForm/index.ts
@@ -1,0 +1,1 @@
+export * from './UserEditForm'
